+++ backend/services/applications.ts
@@ -148,6 +148,14 @@ function search(query: string) {
     return data
 }
 
+// calendar queries
+function interviews() {
+    const data = db.query(
+        `SELECT id, interview_date, title, company FROM applications WHERE interview_date NOT NULL ORDER BY interview_date;`
+    )
+    return data
+}
+
 export {
     getApplications,
     filterApplications,
@@ -158,4 +166,5 @@ export {
     getAAG,
     chart,
     search,
+    interviews,
 }