+++ .github/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## v0.2.0
+
+1. Bugfixes and code improvements
+1. Added search functionality
+1. Added interviews to Dashboard calendar
+
 ## v0.1.0
 
 1. First release! :tada: